# Faster utf8 validation: ~1.14-2x improvement