--- a/benches/read_parquet.rs
+++ b/benches/read_parquet.rs
@@ -58,7 +58,7 @@ fn add_benchmark(c: &mut Criterion) {
         c.bench_function(&a, |b| b.iter(|| read_batch(&buffer, size, 6).unwrap()));
 
         let a = format!("read utf8 emoji 2^{}", i);
-        c.bench_function(&a, |b| b.iter(|| read_batch(&buffer, size, 7).unwrap()));
+        c.bench_function(&a, |b| b.iter(|| read_batch(&buffer, size, 12).unwrap()));
 
         let a = format!("read bool 2^{}", i);
         c.bench_function(&a, |b| b.iter(|| read_batch(&buffer, size, 3).unwrap()));