Two improvements for `try_check_offsets_and_utf8` 
* check bounds afterwards for the last value only as it is the largest value which is guaranteed by monotonicity check (same as `try_check_offsets`).
* run `simdutf8::basic::from_utf8` on the entire buffer instead of individual slices and validate whether the start of the utf8 sequence is correct. I am not 100% sure it is correct,  this is inspired by a recent PR to arrow-rs. It checks whether the first byte is a valid byte to start the sequence.

Change is up to 2x on smaller inputs (that failed to use the simd code before which depends on larger inputs)
```
read utf8 large 2^14    time:   [1.2200 ms 1.2232 ms 1.2262 ms]                                  
                        change: [-13.218% -12.371% -11.592%] (p = 0.00 < 0.05)
                        Performance has improved.

read utf8 emoji 2^14    time:   [130.95 us 131.35 us 131.82 us]                                 
                        change: [-49.840% -49.615% -49.358%] (p = 0.00 < 0.05)
                        Performance has improved.

```