@Dandandan 2022-02-07
> > > 128
> > 
> > 
> > > > run simdutf8::basic::from_utf8 on the entire buffer instead of individual strings and validate whether the start of the utf8 sequence is correct. I am not 100% sure it is correct, this is inspired by a recent PR to arrow-rs. It checks whether the first byte is a valid byte to start the sequence.
> > > 
> > > 
> > > We tried this one before, but sadly we cannot do that optimization. There is no guarantee that a correct entire buffer contains correct utf8 elements. For instance the letter `pi` is: `[207, 128]` which is non-valid utf8 if we split it.
> > 
> > 
> > Yes, I am aware of this. The letter pi with other offsets will not validate as the 128 is not a valid start of the sequence `(128u8 as i8) < -0x40)` evaluates to `true`, so will give an error.
> 
> Yes, I understand the rationale now. So this is sound under the assumption that all multi-byte unicode chars contain values > 127 right? Really interesting observation! 🙌

Yeah, a second 10xxxxxx

> 

Yeah, every second/third/fourth byte always starts with  bits 11 which is what this condition checks for in a clever way.

Also see the encoding table here
https://en.m.wikipedia.org/wiki/UTF-8