@jorgecarleitao 2022-02-07
I PRed a prop test for this to give us extra safety. Really cool optimization!